+++ packages/web/src/components/controls/Input/index.ts
@@ -1,3 +1,4 @@
+import { HTMLAttributes } from 'react'
 import { withStyleProps } from '../../../utils'
 import { StatefulInput } from './Components'
 import { getInputStyles } from './getStyles'
@@ -7,7 +8,8 @@ export * from './Components'
 export * from './getStyles'
 export * from './types'
 
-export const Input = withStyleProps<InputStyleProps, InputProps>(
-  StatefulInput,
-  getInputStyles,
-)
+export const Input = withStyleProps<
+  InputStyleProps,
+  InputProps,
+  HTMLAttributes<HTMLInputElement>
+>(StatefulInput, getInputStyles)