+++ packages/web/src/utils/withStyleProps.tsx
@@ -1,9 +1,9 @@
 import { FunctionComponent } from 'react'
 
 export const withStyleProps =
-  <StyleProps extends {}, Props extends {}>(
-    Component: FunctionComponent<Props>,
-    transformer: (props: Props & StyleProps) => Props,
+  <StyleProps extends {}, Props extends {}, HtmlProps extends {} = {}>(
+    Component: FunctionComponent<Props & HtmlProps>,
+    transformer: (props: Props & StyleProps & HtmlProps) => Props & HtmlProps,
   ) =>
-  (props: Props & StyleProps) =>
+  (props: Props & StyleProps & HtmlProps) =>
     <Component {...transformer(props)} />